pyproject.toml
[project]
name = "pyesi-client"
version = "0.1.0"
description = "EVE Online ESI client library built on pyesi-openapi with caching and utilities"
authors = [{ name = "arxbombus", email = "arxbombus@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyesi-openapi>=0.1.0",
    "pydantic>=2.0.0",
    "pyjwt>=2.10.1",
    "cryptography>=45.0.7",
]

[dependency-groups]
dev = ["pytest>=8.0.0", "pytest-cov>=4.0.0", "pytest-sugar>=1.0.0"]
lint = ["pre-commit>=3.0.0", "ruff>=0.1.0", "pyright>=1.1.405"]

[tool.uv]
python-downloads = "manual"
default-groups = ["dev", "lint"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
